hw_07072023 (21.08.2023 14:45)

+++ servlets_hw_02062023/src/main/java/by/teachmeskills/shop/domain/Cart.java
@@ -10,7 +10,7 @@ import java.util.Map;
 public class Cart {
     private Map<Integer, Product> products;
     @Getter
-    private int totalPrice = 0;
+    private double totalPrice = 0;
 
     public Cart() {
         this.products = new HashMap<>();
